+++ src/media_enrichment/__init__.py
@@ -1,7 +1,7 @@
 """media-enrichment Skill package.
 
-Version: 0.1.0-dev16
+Version: 0.1.0-dev17
 """
 
-__version__ = "0.1.0-dev16"
+__version__ = "0.1.0-dev17"
 SKILL_NAME = "media-enrichment"